streaming/requirements.txt
fastapi
uvicorn[standard]
websockets
kafka-python
opencv-python-headless==4.5.5.64
numpy
